change id columns to bigint

+++ internal/db/models.go
@@ -3,14 +3,14 @@ package db
 import "time"
 
 type GdaxOrderBook struct {
-	ID uint 			`gorm:"primary_key"`
-	MarketID uint
+	ID uint64 			`gorm:"primary_key"`
+	MarketID uint64
 	Time time.Time		`gorm:"primary_key"`
 }
 
 type GdaxOrder struct {
-	ID uint 			`gorm:"primary_key"`
-	OrderbookID uint
+	ID uint64 			`gorm:"primary_key"`
+	OrderbookID uint64
 	Buy bool
 	Rate float64
 	//bitfinex supports giving the total number of sell/buyorders.
@@ -21,14 +21,14 @@ type GdaxOrder struct {
 }
 
 type GdaxMarket struct {
-	ID uint 			`gorm:"primary_key"`
+	ID uint64 			`gorm:"primary_key"`
 	Ticker string		`gorm:"unique_index:idx_market"`
 	Quote string		`gorm:"unique_index:idx_market"`
 }
 
 type GdaxTicker struct {
-	ID  uint 			`gorm:"primary_key"`
-	MarketID uint
+	ID  uint64 			`gorm:"primary_key"`
+	MarketID uint64
 	BestBid float64
 	BestAsk float64
 	Time time.Time		`gorm:"primary_key"`
